First stop toward libevent integration.
All the communications between the priv process and the unpriv
process are now driven by the libevent. This trim a lot of code using
select directly.

I also modified the signals handlers for being handler by the libevent
event loop. This allowed for parameter passing to the signals handlers
and cut down severals global variables. And we all know that global
variables are evil, right ?

The code behave exactly like before, but feel free to review the code as
I might not be aware of the trade-offs of my patch.

By the way I also solved some warning related to 64 to 32 convertions.
It's not a big deal.