Fix the blocking behavior with kqueue backends
The port to libevent induced a block with the kqueue backend because of
the event being edge triggered. Kqueue triggered a write event ways
earlier than expected (compared to epoll) and a bit of code was based on
the assertion that this would not be the case.

However I'm not sure if this patch fixes the issue or the symptoms.