Add the ipv4_prefix_from_netmask function.
Under linux, when we use the netlink interface, the only way to conf the
netmask is to set the prefix (CIDR notation) and the local address at
once. But the current inteface in our code have two functions:
	- one for set the ip
	- one for set the netmask
Therefore, when we use netlink we are compelled to do some little hacks
like setting the addresse twice, but once without netmask and to convert
the netmask to CIDR notation.