Change a bit the cmakelists.txt
Now the cmakelist handle the case where you set the path of libtuntap
and tapcfg by hand. There is no more probleme with include being out of
/usr/include or /usr/local/include.

If the cmake don't find libtuntap or tapcfg you have to set the
variables by hand with cmake-gui, ccmake or by editing the
CMakeCache.txt.