add: udp port announce
This is the very step toward DTLS integration. I refactored most of the
UDP code and make it a little bit more like a standalone module.

Now, the peer exchange their UDP port via the meta-connexion, with the
udp_port:<port> command.

Also, the UDP port is now chose randomly.