subset: initial commit
Create a new header and source files containing all the (future) tricks
in C, related to simplicity and correctness of the code.

The fist citizen of this subset is "tnt_new". It's a macro function
taking a type as parameter a type, mallocing it, memseting it to zero,
and returning a "type *".

This macro gives two advantages:
	- your memory is always zeroed (sorry for the perf)
	- it returns a pointer of the right type, catching more error at
	  compile time when you malloc a memory of some type using the
	  size of another type

Well, I know that some of us are C nazi, and would like to see this kind
of C++ish burned.

But, Fuck me, right ?