Update our cmake modules, the FindXXX stuff
Remove the need for LibFindMacros.cmake. We ask for 2.8, so we don't
need compatibility goo with 2.6.

Also add a nice STATUS message when a package is found without
pkg-config, such as manually installed or system provided stuff.

Except for the licence header, our FindXXX modules are compliant with
the cmake formating rules.